refactor: use single chacha + aes impl
