+++ package.json
@@ -4,6 +4,20 @@
 	"description": "WebCrypto Based Cross Platform TLS",
 	"main": "lib/index",
 	"type": "module",
+	"exports": {
+		".": {
+			"default": "./lib/index.js",
+			"types": "./lib/index.d.ts"
+		},
+		"./webcrypto": {
+			"default": "./lib/crypto/webcrypto.js",
+			"types": "./lib/crypto/webcrypto.d.ts"
+		},
+		"./purejs-crypto": {
+			"default": "./lib/crypto/purejs.js",
+			"types": "./lib/crypto/purejs.d.ts"
+		}
+	},
 	"scripts": {
 		"build": "npm exec tsc -- -p tsconfig.build.json",
 		"run:tsc": "node --experimental-strip-types",
@@ -48,7 +62,6 @@
 		"@peculiar/asn1-ecc": "^2.3.14",
 		"@peculiar/asn1-schema": "^2.3.13",
 		"@peculiar/x509": "^1.12.3",
-		"@stablelib/chacha20poly1305": "^1.0.1",
 		"micro-rsa-dsa-dh": "^0.1.0"
 	},
 	"devDependencies": {
