+++ src/tests/load-purejs-crypto.ts
@@ -1,4 +1,4 @@
 import { setCryptoImplementation } from '../crypto/index.ts'
-import { pureJsCrypto } from '../crypto/pure.ts'
+import { pureJsCrypto } from '../crypto/pure-js.ts'
 
 setCryptoImplementation(pureJsCrypto)
